Move AWS SDK credential bridge integration test to the existing AWS SDK integration test run
